+++ packages/synapse-interface/components/Activity/Activity.tsx
@@ -2,20 +2,17 @@ import React, { useMemo } from 'react'
 import _ from 'lodash'
 import Fuse from 'fuse.js'
 import { useAccount } from 'wagmi'
-import { Address } from 'viem'
+import { type Address } from 'viem'
+import { type Chain } from '@/utils/types'
 import { useTransactionsState } from '@/slices/transactions/hooks'
 import { usePortfolioState } from '@/slices/portfolio/hooks'
 import { BridgeTransaction } from '@/slices/api/generated'
 import { CHAINS_BY_ID } from '@/constants/chains'
-import { Chain, Token } from '@/utils/types'
 import { tokenAddressToToken } from '@/constants/tokens'
 import { TransactionsState } from '@/slices/transactions/reducer'
 import { PortfolioState } from '@/slices/portfolio/reducer'
-import {
-  Transaction,
-  TransactionType,
-} from '../Portfolio/Transaction/Transaction'
-import { UserExplorerLink } from '../Portfolio/Transaction/components/TransactionExplorerLink'
+import { Transaction, TransactionType } from './Transaction/Transaction'
+import { UserExplorerLink } from './Transaction/components/TransactionExplorerLink'
 import { NoSearchResultsContent } from '../Portfolio/components/NoSearchResultContent'
 import { checkTransactionsExist } from '@/utils/checkTransactionsExist'
