+++ packages/synapse-interface/components/Activity/Transaction/components/TransactionPayloadDetail.tsx
