+++ packages/synapse-interface/components/Portfolio/hooks/useSearchInputStatus.ts
