+++ packages/synapse-interface/components/_Transaction/_Transaction.tsx
@@ -5,7 +5,7 @@ import { getExplorerAddressLink } from './helpers/getExplorerAddressLink'
 import { useBridgeTxStatus } from './helpers/useBridgeTxStatus'
 import { isNull } from 'lodash'
 import { removeTransaction } from '@/slices/_transactions/reducer'
-import { TransactionPayloadDetail } from '../Portfolio/Transaction/components/TransactionPayloadDetail'
+import { TransactionPayloadDetail } from '../Activity/Transaction/components/TransactionPayloadDetail'
 import { Chain, Token } from '@/utils/types'
 import { TimeRemaining } from './components/TimeRemaining'
 import { calculateEstimatedTimeStatus } from './helpers/calculateEstimatedTimeStatus'
