+++ packages/synapse-interface/utils/hooks/useIsFocused.ts
