+++ packages/synapse-interface/utils/hooks/useIsMounted.ts
@@ -1,8 +1,5 @@
 import { useState, useEffect } from 'react'
 
-/**
- * Checks when app has mounted
- */
 export const useIsMounted = (): boolean => {
   const [mounted, setMounted] = useState<boolean>(false)
